adjusted controller, services, models, dtos, ajusted queries, and got rid of non-used code

+++ tenmo-server/src/main/java/com/techelevator/tenmo/service/AccountService.java
@@ -1,14 +1,14 @@
 package com.techelevator.tenmo.service;
 
-import com.techelevator.tenmo.model.Account;
+import com.techelevator.tenmo.entity.Account;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface AccountService {
-    Optional<Account> getAccountById(int accountId);
+    Optional<Account> getAccountById(Long accountId);
 
-    Optional<Account> getAccountByUserId(int userId);
+    Optional<Account> getAccountByUserId(Long userId);
 
     List<Account> getAccounts();
 }
